+++ application/controllers/Kriteria.php
@@ -10,15 +10,6 @@
             $this->load->library('pagination');
             $this->load->library('form_validation');
             $this->load->model('Kriteria_model');
-
-            if ($this->session->userdata('id_user_level') != "1") {
-            ?>
-				<script type="text/javascript">
-                    alert('Anda tidak berhak mengakses halaman ini!');
-                    window.location='<?php echo base_url("Login/home"); ?>'
-                </script>
-            <?php
-			}
         }
 
         public function index()
@@ -31,6 +22,14 @@
         //menampilkan view create
         public function create()
         {
+            if ($this->session->userdata('id_user_level') != "1") {
+                ?>
+                    <script type="text/javascript">
+                        alert('Anda tidak berhak mengakses halaman ini!');
+                        window.location='<?php echo base_url("Login/home"); ?>'
+                    </script>
+                <?php
+            }
 			$data['page'] = "Kriteria";
             $this->load->view('kriteria/create', $data);
         }
@@ -60,7 +59,7 @@
                     }
                 } else {
                     $this->session->set_flashdata('message', '<div class="alert alert-danger" role="alert">Data gagal disimpan!</div>');
-                    redirect('Kriteria');
+                    redirect('Kriteria/create');
 
                 }
             
@@ -69,6 +68,14 @@
 
         public function edit($id_kriteria)
         {
+            if ($this->session->userdata('id_user_level') != "1") {
+                ?>
+                    <script type="text/javascript">
+                        alert('Anda tidak berhak mengakses halaman ini!');
+                        window.location='<?php echo base_url("Login/home"); ?>'
+                    </script>
+                <?php
+            }
             $data['page'] = "Kriteria";
 			$data['kriteria'] = $this->Kriteria_model->show($id_kriteria);
             $this->load->view('kriteria/edit', $data);
@@ -76,37 +83,58 @@
     
         public function update($id_kriteria)
         {
-            // TODO: implementasi update data berdasarkan $id_kriteria
-            $id_kriteria = $this->input->post('id_kriteria');
-            $data = array(
-                'keterangan' => $this->input->post('keterangan'),
-                'kode_kriteria' => $this->input->post('kode_kriteria'),
-                'bobot' => $this->input->post('bobot'),
-                'jenis' => $this->input->post('jenis')
-            );
+            if ($this->session->userdata('id_user_level') == "1") {
+                // TODO: implementasi update data berdasarkan $id_kriteria
+                $id_kriteria = $this->input->post('id_kriteria');
+                $data = array(
+                    'keterangan' => $this->input->post('keterangan'),
+                    'kode_kriteria' => $this->input->post('kode_kriteria'),
+                    'bobot' => $this->input->post('bobot'),
+                    'jenis' => $this->input->post('jenis')
+                );
 
-            // Validasi update data  (https://stackoverflow.com/questions/27621250/is-unique-in-codeigniter-for-edit-function) Ellix4u's solution
-            $id = $this->uri->segment(3);
-            $this->form_validation->set_rules('keterangan', 'Keterangan', 'required|edit_unique[kriteria.keterangan.id_kriteria.'.$id.']');
-            $this->form_validation->set_rules('kode_kriteria', 'Kode Kriteria', 'required|edit_unique[kriteria.kode_kriteria.id_kriteria.'.$id.']');
-            if ($this->form_validation->run() != false) {
-                $this->Kriteria_model->update($id_kriteria, $data);
-                $this->session->set_flashdata('message', '<div class="alert alert-success" role="alert">Data berhasil diupdate!</div>');
-                redirect('Kriteria');
-            } else {
-                $this->session->set_flashdata('message', '<div class="alert alert-danger" role="alert">Data gagal diupdate!</div>');
-                redirect('Kriteria');
+                // Validasi update data  (https://stackoverflow.com/questions/27621250/is-unique-in-codeigniter-for-edit-function) Ellix4u's solution
+                $id = $this->uri->segment(3);
+                $this->form_validation->set_rules('keterangan', 'Keterangan', 'required|edit_unique[kriteria.keterangan.id_kriteria.'.$id.']');
+                $this->form_validation->set_rules('kode_kriteria', 'Kode Kriteria', 'required|edit_unique[kriteria.kode_kriteria.id_kriteria.'.$id.']');
+                if ($this->form_validation->run() != false) {
+                    $this->Kriteria_model->update($id_kriteria, $data);
+                    $this->session->set_flashdata('message', '<div class="alert alert-success" role="alert">Data berhasil diupdate!</div>');
+                    redirect('Kriteria');
+                } else {
+                    $this->session->set_flashdata('message', '<div class="alert alert-danger" role="alert">Data gagal diupdate!</div>');
+                    redirect('Kriteria/edit/'.$id_kriteria);
+                }
+                // $this->Kriteria_model->update($id_kriteria, $data);
+                // $this->session->set_flashdata('message', '<div class="alert alert-success" role="alert">Data berhasil diupdate!</div>');
+                // redirect('Kriteria');
             }
-            // $this->Kriteria_model->update($id_kriteria, $data);
-            // $this->session->set_flashdata('message', '<div class="alert alert-success" role="alert">Data berhasil diupdate!</div>');
-			// redirect('Kriteria');
+            
+            if ($this->session->userdata('id_user_level') != "1") {
+                ?>
+                    <script type="text/javascript">
+                        alert('Anda tidak berhak mengakses halaman ini!');
+                        window.location='<?php echo base_url("Login/home"); ?>'
+                    </script>
+                <?php
+            }
         }
     
         public function destroy($id_kriteria)
         {
-            $this->Kriteria_model->delete($id_kriteria);
-            $this->session->set_flashdata('message', '<div class="alert alert-success" role="alert">Data berhasil dihapus!</div>');
-			redirect('Kriteria');
+            if ($this->session->userdata('id_user_level') == "1") {
+                $this->Kriteria_model->delete($id_kriteria);
+                $this->session->set_flashdata('message', '<div class="alert alert-success" role="alert">Data berhasil dihapus!</div>');
+                redirect('Kriteria');
+            }
+            if ($this->session->userdata('id_user_level') != "1") {
+                ?>
+                    <script type="text/javascript">
+                        alert('Anda tidak berhak mengakses halaman ini!');
+                        window.location='<?php echo base_url("Login/home"); ?>'
+                    </script>
+                <?php
+            }
         }
     
     }
